+++ R/as_draws_array.R
@@ -96,6 +96,7 @@ as_draws_array.draws_rvars <- function(x, ...) {
   x <- check_variables_are_numeric(
     x, to = "draws_array", is_non_numeric = is_rvar_factor, convert = FALSE
   )
+  x <- promote_rvar_weights_to_variable(x)
 
   # cbind discards class information when applied to vectors, which converts
   # the underlying factors to numeric