+++ R/pit.R
@@ -80,7 +80,7 @@ pit.draws_matrix <- function(x, y, weights = NULL, log = FALSE, ...) {
   y <- validate_y(y, x)
   if (!is.null(weights)) {
     weights <- sapply(seq_len(nvariables(x)), \(var_idx) {
-      validate_weights(weights[, var_idx], x[, var_idx], log)
+      validate_weights(weights[, var_idx], ndraws(x), log)
     })
     weights <- normalize_log_weights(weights)
   }